+++ eelbrain/pipeline.py
@@ -1,5 +1,5 @@
 from ._experiment.mne_experiment import MneExperiment
-from ._experiment.preprocessing import RawSource, RawFilter, RawICA, RawMaxwell, RawReReference
+from ._experiment.preprocessing import RawSource, RawFilter, RawICA, RawMaxwell, RawReReference, RawApplyICA
 from ._experiment.epochs import PrimaryEpoch, SecondaryEpoch, SuperEpoch
 from ._experiment.groups import Group, SubGroup
 from ._experiment.parc import CombinationParc, FreeSurferParc, FSAverageParc, SeededParc, IndividualSeededParc